+++ src/gui.js
@@ -87,11 +87,11 @@ HatBlockMorph, ZOOM*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2026-August-12';
+modules.gui = '2026-August-13';
 
 // Declarations
 
-var SnapVersion = '12.1.0-rc2';
+var SnapVersion = '12.1.0-rc3';
 
 var IDE_Morph;
 var ProjectDialogMorph;